Add Int64 duration formats for int64-sized wire types

+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.ClientModel/test/Providers/MrwSerializationTypeDefinitions/TestData/DeserializationTests/TestDeserializationOfDurationMillisecondsIntegerWireType(Integer).cs
@@ -24,7 +24,7 @@ namespace Sample.Models
             {
                 if (prop.NameEquals("audio_end_ms"u8))
                 {
-                    audioEndMs = global::System.TimeSpan.FromMilliseconds(prop.Value.GetInt32());
+                    audioEndMs = global::System.TimeSpan.FromMilliseconds(prop.Value.GetInt64());
                     continue;
                 }
                 if ((options.Format != "W"))
